pyproject.toml
[tool.poetry]
name = "drfecommerce"
version = "0.1.0"
description = ""
authors = ["AnishMachamasi <anishmachamasi2262@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
djangorestframework = "^3.15.2"
pygments = "^2.18.0"
asgiref = "^3.8.1"
sqlparse = "^0.5.1"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.2"
django-mptt = "^0.16.0"
pytest = "^8.3.3"
pytest-factoryboy = "^2.7.0"
pytest-django = "^4.9.0"
black = "^24.8.0"
coverage = "^7.6.1"
pytest-cov = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
